修改类型

+++ frameworks/js/napi/session/include/napi_common_define.h
@@ -69,7 +69,7 @@ struct NativeClass {
     napi_callback constructor;
     napi_ref *constructorRef;
     napi_property_descriptor *desc;
-    int32_t descSize = 0;
+    size_t descSize = 0;
 };
 
 template<typename T>
